Split `main.rs` into five separate source files for better organization
This is just the beginning of a big refactor.
So far I've been prototyping all the code in `main.rs` with little
consideration for the greater architecture and code organization, since
I still needed to figure out what libraries to even use and how to best
incorporate them.

With almost all core functionality of the app in place,
now is a good moment to refactor and reorganize the code.