Replace custom AppError macros with utility functions instead
Instead of being a thin wrapper around anyhow::Error, AppError now
contains a StatusCode and String-message which will be turned into the
appropriate response.

For convenience, the From<...> implementation for anyhow::error is kept,
but the actual method immediately extracts the String and stores it as
the AppError's `message`.

The custom bail! and ensure! macros are now replaced with new functions
on AppError that quickly construct it or Result<T, AppError>::Err of it:
::new, ::new500 and ::err