Fix a whole bunch of issues spotted by Lighthouse
Fixes:
- insufficient contrasts on various elements
- incorrect heading hierarchies (h4 below h2, etc.)
- missing labels on mobile icon-only links in header

Also fixes:
Previously the download-page performed a lot of templating in the
client. This has now been moved to the server and some elements are now
making use of 'visibility: hidden;' in order to not be visible but still
consume layout-space.
All of this was done to help the download-page reach a 100 Lighthouse
Performance score because the previous method caused big Cumulative
Layout Shift once the client-side decryption was done and the
download-box was suddenly dropped onto the page.