Enforce 2GiB filesize limit and catch frontend encryption errors
Testing revealed that the WebCrypto-API refuses to encrypt messages
larger than 2GiB, thereby effectively setting the maximum filesize limit
to just that, 2GiB. This commit updates the configuration wizard to
account for this fact. Additionally, should a larger maximum filesize
already be configured, it will be clamped to 2GiB during application
startup. There's no point in configuring a larger filesize on the
backend if a user cannot possibly upload files that large.

This commit also catches encryption errors on the frontend and displays
an appropriate error message to the user should they occur.