Remove unsubsetted fonts from the repository
Even though Git LFS is the better solution for storing large, binary
assets in the repository, its use is quite restricted on GitHub's
free plan. Only 1GB of free bandwidth is provided per month, which
means a repository with 5MB of Git LFS assets in it can only be
cloned 200 times in a month before hitting quota limits.

This commit does not move the repository off of Git LFS entirely
(although I might consider that move in the future), but instead removes
the unsubsetted font files from the `font/`-directory.
These were only used during developement and are not at all required to
build the application from source.