Use deterministic IVs for AES-GCM instead of randomly generated ones
This changes isn't strictly necessary, but aligns FerriShare with the
recommended best practice for IV construction when using AES-GCM.

NIST states that randomly chosen IVs with AES-GCM are safe for up to
2^32 messages and FerriShare only generates two messages, putting
the chance of an IV collision at 1 in 2^96.

However, there is no need for FerriShare to randomly generate IVs.
Since only two messages are generted for each randomly chosen key,
the IVs can be chosen deterministically and the chance of an
IV collision can be reduced to 0.