Only display the author's name in the footer if running in demo mode
This removes the "built with Rust and <3 by Tobias Marschner" from the
footer unless FerriShare is running in demo mode (as is the case on
ferrishare-demo.tobiasm.dev).

After some consideration, I realized that my name in the footer on
other people's instances might suggest to a visitor that I am somehow
involved in operating or maintaining that instance, which is obviously
not the case.
To proactively prevent confusion, I decided to remove my name from the
footer in normal installations. The link to the repository and the
version tag remain in place.